feat(daemon): Phase 5 — pressure signal + cascade demote + USN background priority (5.3, 5.6, 5.7, 5.10)
Closes the four remaining Phase 5 tasks from the local-only memory-tiering plan, queued behind PR #104.

5.3 PressureSignal trait + PlatformPressureSignal (Mac/Linux never-fires; Win watcher deferred). 5.6 spawn_pressure_subscriber + cascade_demote_one_step. 5.7 BackgroundIoPriority trait + RAII BackgroundIoScope wrapping each per-letter spawn_blocking in refresh_usn_for_warm_shards. 5.10 LRU-cascade + USN-refresh tests.

Mechanical: split 3008-line tests.rs into tests/ module dir for the four new injection-test files.

Mac gate: nextest workspace 1506/1506 + daemon 177/177 ×5 idempotency, lint-prod clean, check-windows clean.
